Added line break to Combat buttons
Added a line break to each of the combat buttons, Attack, Potion, Run, to address misalignment of run button text due to smaller icon image not big enough to cause wrapping of button text.

Minor edits: Removed a span element from the Potion button that is no longer used. Previously used to hold current number of potions. This is deprecated since number of potions is now displayed in the stats banner.